Changed Controllers, some models, and added some Requests

+++ routes/web.php
@@ -6,6 +6,8 @@ use App\Http\Controllers\UserController;
 use App\Http\Controllers\CategoryController;
 use App\Http\Controllers\CountryController;
 use App\Http\Controllers\StateController;
+use App\Http\Controllers\StatusController;
+use App\Http\Controllers\RoleController;
 use App\Http\Controllers\EventController;
 use App\Http\Controllers\RegistrationController;
 
@@ -21,6 +23,10 @@ Route::controller(AuthController::class)->group(function () {
 Route::controller(UserController::class)->group(function () {
     // ADMIN
     Route::get("/api/v1/users", "index")->middleware(["auth:sanctum", "abilities:ADMIN"]);
+    Route::get("/api/v1/users/{id}", "show")->middleware(["auth:sanctum", "abilities:ADMIN"]);
+    Route::put("/api/v1/users/{id}", "update")->middleware(["auth:sanctum", "abilities:ADMIN"]);
+    Route::patch("/api/v1/users/{id}", "restore")->middleware(["auth:sanctum", "abilities:ADMIN"]);
+    Route::delete("/api/v1/users/{id}", "destroy")->middleware(["auth:sanctum", "abilities:ADMIN"]);
     // PUBLIC
     Route::post("/api/v1/users", "store");
 });
@@ -44,10 +50,9 @@ Route::controller(CountryController::class)->middleware(["auth:sanctum", "abilit
         Route::get("/api/v1/countries/{id}", "show");
         Route::put("/api/v1/countries/{id}", "update");
         Route::patch("/api/v1/countries/{id}", "restore");
-        Route::delete("/api/v1/categories/{id}", "destroy");
+        Route::delete("/api/v1/countries/{id}", "destroy");
 });
 
-// ADMIN
 Route::controller(StateController::class)->group(function () {
     // PUBLIC
     Route::get("/api/v1/states", "index");
@@ -59,20 +64,43 @@ Route::controller(StateController::class)->group(function () {
     Route::delete("/api/v1/states/{id}", "destroy")->middleware(["auth:sanctum", "abilities:ADMIN"]);
 });
 
-Route::controller(EventController::class)->middleware(["auth:sanctum"])
+// ADMIN
+Route::controller(StatusController::class)->middleware(["auth:sanctum", "abilities:ADMIN"])
     ->group(function () {
-        Route::get("/api/v1/events", "index");
-        Route::post("/api/v1/events", "store")->middleware(["ability:ORGANIZER,ADMIN"]);
-        Route::get("/api/v1/events/{id}", "show");
-        Route::put("/api/v1/events/{id}", "update")->middleware(["abilities:ADMIN"]);
-        Route::patch("/api/v1/events/{id}", "restore")->middleware(["abilities:ADMIN"]);
-        Route::delete("/api/v1/events/{id}", "destroy")->middleware(["abilities:ADMIN"]);
+        Route::get("/api/v1/status", "index");
+        Route::post("/api/v1/status", "store");
+        Route::get("/api/v1/status/{id}", "show");
+        Route::put("/api/v1/status/{id}", "update");
+        Route::patch("/api/v1/status/{id}", "restore");
+        Route::delete("/api/v1/status/{id}", "destroy");
+});
+
+// ADMIN
+Route::controller(RoleController::class)->middleware(["auth:sanctum", "abilities:ADMIN"])
+    ->group(function () {
+        Route::get("/api/v1/roles", "index");
+        Route::post("/api/v1/roles", "store");
+        Route::get("/api/v1/roles/{id}", "show");
+        Route::put("/api/v1/roles/{id}", "update");
+        Route::patch("/api/v1/roles/{id}", "restore");
+        Route::delete("/api/v1/roles/{id}", "destroy");
+});
+
+Route::controller(EventController::class)->group(function () {
+    Route::get("/api/v1/events", "index");
+    Route::post("/api/v1/events", "store")->middleware(["auth:sanctum", "abilities:ADMIN"]);
+    Route::get("/api/v1/events/{id}", "show");
+    Route::put("/api/v1/events/{id}", "update")->middleware(["auth:sanctum", "abilities:ADMIN"]);
+    Route::patch("/api/v1/events/{id}", "restore")->middleware(["auth:sanctum", "abilities:ADMIN"]);
+    Route::delete("/api/v1/events/{id}", "destroy")->middleware(["auth:sanctum", "abilities:ADMIN"]);
 });
 
 Route::controller(RegistrationController::class)->middleware(["auth:sanctum"])
     ->group(function () {
-        Route::get("/api/v1/registrations", "index");
+        Route::get("/api/v1/registrations", "index")->middleware(["abilities:ADMIN"]);
         Route::post("/api/v1/registrations", "store");
-        Route::delete("/api/v1/registrations/{registration}", "destroy");
+        Route::put("/api/v1/registrations/{id}", "update")->middleware(["abilities:ADMIN"]);
+        Route::patch("/api/v1/registrations/{id}", "restore")->middleware(["abilities:ADMIN"]);
+        Route::delete("/api/v1/registrations/{id}", "destroy")->middleware(["abilities:ADMIN"]);
 });
